Moved show logo logic function to log file
