+++ src/cromshell/utilities/http_utils.py
@@ -4,6 +4,7 @@ import sys
 import requests
 
 from cromshell.utilities import io_utils
+from cromshell import log
 
 LOGGER = logging.getLogger(__name__)
 
@@ -15,13 +16,14 @@ def assert_can_communicate_with_server(config):
         request_out = requests.get(f"{config.cromwell_server}{config.api_string}backends", timeout=5)
     except ConnectionError:
         LOGGER.error("Failed to connect to %s", config.cromwell_server)
-        sys.exit(1)
+        raise Exception("Failed to connect to %s", config.cromwell_server)
     except requests.exceptions.RequestException:
-        raise Exception(f"Failed to connect to {config.cromwell_server}") from None
+        LOGGER.error("Failed to connect to %s", config.cromwell_server)
+        raise Exception(f"Failed to connect to {config.cromwell_server}")
 
     if b"supportedBackends" not in request_out.content:
-        io_utils.turtle_dead(config.show_logo)
+        log.display_logo(io_utils.dead_turtle)
         LOGGER.error(
             "Error: Cannot communicate with Cromwell server: %s", config.cromwell_server
         )
-        sys.exit(1)
+        raise Exception("Error: Cannot communicate with Cromwell server: %s", config.cromwell_server)
